requirements.txt
certifi==2018.10.15
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.2
django-discover-runner==1.0
django-filter==2.0.0
django-rest-swagger==2.2.0
django-role-permissions==2.2.0
djangorestframework==3.8.2
djangorestframework-simplejwt==3.2.3
gunicorn==19.9.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.0
mysqlclient==1.3.13
openapi-codec==1.3.2
Pygments==2.3.0
PyJWT==1.6.4
python-decouple==3.1
pytz==2018.5
requests==2.20.0
simplejson==3.16.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.24
snownlp==0.12.3
